refactor(front): uniformized folder names for pages
